requirements.txt
Flask==2.3.3
beautifulsoup4==4.12.2
pywebview==4.3.3
chardet==5.2.0
python-docx==0.8.11
docx2txt==0.8
markdown==3.4.4
pyinstaller==5.13.2